e2g-demo/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "e2g-demo"
version = "1.0.0"
description = "Federated Learning with E2G and Flower using AWS and NextFlow"
license = "Apache-2.0"
dependencies = [
    "flwr==1.14.0",
    "boto3>=1.26.0",
    "botocore>=1.29.0",
    "tensorflow-cpu==2.15.1",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.flwr.app]
publisher = "flwrlabs"

[tool.flwr.app.components]
serverapp = "e2gflower.server_app:app"
clientapp = "e2gflower.client_app:app"

[tool.flwr.app.config]
num-server-rounds = 1 # 5
fraction-fit = 1.0
batch-size = 128

[tool.flwr.federations]
default = "e2gdemo"

[tool.flwr.federations.e2gdemo]
address = "10.0.175.85:9093" # "localhost:9093"
insecure = true

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint]
extend-select = ["UP", "RUF", "I", "D", "E501"]
ignore = ['UP036']

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D100", "D103", "D104"]
"__main__.py" = ["D100"]
"__init__.py" = ["D104"]
